Target to generate VirtualBox image
The kernel command line can be different when using the disk installer.
This allows the console to go to the screen for the disk installer while
still going to the serial port for normal boots.

Remove a check for VBoxMange being installed.

Adds some somewhat more clear instructions on completion or failure.

Change-Id: Ic221d6c75b3109b2ec95a6c6a650c008f1dfef87