vbox_x86 builds: force-set the vdi UUIDs
The VDI UUID is used by VirtualBox to make sure that a disk is not in use by
more than one VM at a time.  This is a nuisance when you have to rebuild the
image constantly, as you have to manually release and re-add the VDI with a new
UUID each time before you run the VM.

So for now, set the UUID to all As for /dev/sda and all Bs for /dev/sdb to make
life easier.

Change-Id: I3e36331c25b6f47c2d466556f906b4a11ef53e74
Signed-off-by: Iliyan Malchev <malchev@google.com>